Implement different insert methods
Now we are ready to implement insert functionality. The usual suspects
are row-by-row INSERT, batched INSERT and COPY. Let's first add two more
command-line parameters for the final number of rows and the batch size.

We introduce here a struct `DbUser` with a `TestUser` variable which
holds some constant values. At this phase we will insert the same data
over and over again. That's enough for our goal.

Now we add three functions `InsertSimple()`, `InsertBatch()` and
`InsertCopy()`.